+++ controllers/event/event.controller.js
@@ -1,5 +1,5 @@
-const contestService = require("../services/event/contest.service.js");
-const logger = require("../logger");
+const contestService = require("../../services/event/contest.service.js");
+const logger = require("../../logger");
 
 exports.getAllContests = async (req, res, next) => {
   try {
